# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MemorialSpace API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations:
    get:
      operationId: listOrganizations
      tags: [organizations]
      summary: List all cemetery organizations
      responses:
        "200":
          description: List of organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
    post:
      operationId: createOrganization
      tags: [organizations]
      summary: Create a new cemetery organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationBody"
      responses:
        "201":
          description: Created organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /organizations/{id}:
    get:
      operationId: getOrganization
      tags: [organizations]
      summary: Get organization by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    put:
      operationId: updateOrganization
      tags: [organizations]
      summary: Update an organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationBody"
      responses:
        "200":
          description: Updated organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    delete:
      operationId: deleteOrganization
      tags: [organizations]
      summary: Delete an organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── USERS ───────────────────────────────────────────────────────────────────
...
  /memorials:
    get:
      operationId: listMemorials
      tags: [memorials]
      summary: List memorial pages
      parameters:
        - name: organizationId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of memorials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Memorial"
    post:
      operationId: createMemorial
      tags: [memorials]
      summary: Create a memorial page
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMemorialBody"
      responses:
        "201":
          description: Created memorial
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memorial"
  /memorials/{id}:
    get:
      operationId: getMemorial
      tags: [memorials]
      summary: Get memorial page by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Memorial
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memorial"
    put:
      operationId: updateMemorial
      tags: [memorials]
      summary: Update a memorial page
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMemorialBody"
      responses:
        "200":
          description: Updated memorial
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memorial"
  /memorials/{id}/tributes:
    get:
      operationId: listTributes
      tags: [memorials]
      summary: List tribute messages for a memorial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of tributes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tribute"
    post:
      operationId: createTribute
      tags: [memorials]
      summary: Add a tribute message to a memorial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTributeBody"
      responses:
        "201":
          description: Created tribute
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tribute"

  # ─── WORK ORDERS ─────────────────────────────────────────────────────────────
...
  /marketplace/products:
    get:
      operationId: listProducts
      tags: [marketplace]
      summary: List marketplace products
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [flowers, urns, services, other]
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [marketplace]
      summary: Create a marketplace product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /marketplace/products/{id}:
    get:
      operationId: getProduct
      tags: [marketplace]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [marketplace]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [marketplace]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /marketplace/orders:
    get:
      operationId: listOrders
      tags: [marketplace]
      summary: List marketplace orders
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [marketplace]
      summary: Place a marketplace order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /marketplace/orders/{id}:
    get:
      operationId: getOrder
      tags: [marketplace]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ─── DASHBOARD ───────────────────────────────────────────────────────────────
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments
      parameters:
        - name: organizationId
          in: query
          schema:
            type: integer
        - name: invoiceId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Record a payment against an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Created payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    delete:
      operationId: deletePayment
      tags: [payments]
      summary: Delete a payment (recomputes parent invoice)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: organizationId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── ACCOUNTING SUMMARY ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    Organization:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        cemeteryType:
          type: string
          enum: [church, private, pet, municipality, columbarium]
          description: Type of cemetery chosen at signup. Drives module visibility (e.g. enables the Columbarium module).
        featuresColumbarium:
          type: boolean
          description: True if this cemetery operates a columbarium. Always true when cemeteryType=columbarium; opt-in for cemeteryType=church.
        address:
          type: string
        city:
          type: string
        country:
          type: string
        phone:
          type: string
        email:
          type: string
        website:
          type: string
        logoUrl:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        totalPlots:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - cemeteryType
        - featuresColumbarium
        - createdAt
    CreateOrganizationBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        cemeteryType:
          type: string
          enum: [church, private, pet, municipality, columbarium]
        featuresColumbarium:
          type: boolean
        address:
          type: string
        city:
          type: string
        country:
          type: string
        phone:
          type: string
        email:
          type: string
        website:
          type: string
        logoUrl:
          type: string
        latitude:
          type: number
        longitude:
          type: number
      required:
        - name
        - slug
...
    Memorial:
      type: object
      properties:
        id:
          type: integer
        burialId:
          type: integer
        organizationId:
          type: integer
        title:
          type: string
        biography:
          type: string
        photos:
          type: array
          items:
            type: string
        isPublic:
          type: boolean
        viewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - burialId
        - organizationId
        - title
        - createdAt
    CreateMemorialBody:
      type: object
      properties:
        burialId:
          type: integer
        organizationId:
          type: integer
        title:
          type: string
        biography:
          type: string
        photos:
          type: array
          items:
            type: string
        isPublic:
          type: boolean
      required:
        - burialId
        - organizationId
        - title
    Tribute:
      type: object
      properties:
        id:
          type: integer
        memorialId:
          type: integer
        authorName:
          type: string
        authorEmail:
          type: string
        message:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - memorialId
        - authorName
        - message
        - createdAt
    CreateTributeBody:
      type: object
      properties:
        authorName:
          type: string
        authorEmail:
          type: string
        message:
          type: string
      required:
        - authorName
        - message
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [flowers, urns, services, other]
        price:
          type: number
        imageUrl:
          type: string
        inStock:
          type: boolean
        stockCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - price
        - inStock
        - createdAt
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [flowers, urns, services, other]
        price:
          type: number
        imageUrl:
          type: string
        inStock:
          type: boolean
        stockCount:
          type: integer
      required:
        - name
        - category
        - price
    Order:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        totalAmount:
          type: number
        status:
          type: string
          enum: [pending, paid, processing, shipped, delivered, cancelled]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - customerName
        - items
        - totalAmount
        - status
        - createdAt
    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
      required:
        - productId
        - quantity
        - unitPrice
    CreateOrderBody:
      type: object
      properties:
        customerId:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        totalAmount:
          type: number
      required:
        - customerName
        - items
        - totalAmount
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        organizationId:
          type: integer
        invoiceId:
          type: integer
        amount:
          type: number
        paymentDate:
          type: string
        method:
          type: string
          enum: [cash, check, card, bank_transfer, other]
        reference:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - organizationId
        - invoiceId
        - amount
        - paymentDate
        - method
        - createdAt
    CreatePaymentBody:
      type: object
      properties:
        organizationId:
          type: integer
        invoiceId:
          type: integer
        amount:
          type: number
        paymentDate:
          type: string
        method:
          type: string
          enum: [cash, check, card, bank_transfer, other]
        reference:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - organizationId
        - invoiceId
        - amount
        - paymentDate
        - method
...